+++ GameServer/Response.java
@@ -1,0 +1,7 @@
+package GameServer;
+
+public class Response {
+
+
+
+}
