+++ client/html/game.html
@@ -9,13 +9,13 @@
     </head>
     <body>
         <header id="navbar">
-            <a href="index.html">Home</a>
-            <a href="invite.html">Invite</a>
-            <a href="user.html">My Profile</a>
-            <a href="game.html">Game</a>
-            <form id="logout" method="post">
+            <a href="index.html"><button style="background-color: #add8e6">&#127968; Home</button></a>
+            <a href="invite.html"><button style="background-color: #add8e6">&#9742; Invite</button></a>
+            <a href="user.html"><button style="background-color: #add8e6">&#128100; My Profile</button></a>
+            <a href="game.html"><button style="background-color: #fdfd96">&#127922; Game</button></a>
+            <form id="logout" method="post" style="display:inline;">
                 <input name="action" value="logout" hidden>
-                <button type="submit">Logout</button>
+                <button type="submit" style="display:inline; background-color: #add8e6">&#128272; Logout</button>
             </form>
         </header>
         <h1>Game page</h1>
