+++ client/html/index.html
@@ -9,7 +9,7 @@
     </head>
     <body>
         <header id="navbar">
-            <a href="index.html"><button style="background-color: #add8e6">&#127968; Home</button></a>
+            <a href="index.html"><button style="background-color: #fdfd96">&#127968; Home</button></a>
             <a href="invite.html"><button style="background-color: #add8e6">&#9742; Invite</button></a>
             <a href="user.html"><button style="background-color: #add8e6">&#128100; My Profile</button></a>
             <a href="game.html"><button style="background-color: #add8e6">&#127922; Game</button></a>
